Remove root request
The root request `/` works fine locally because we are not going through an AWS Gateway. But when talking to an AWS environment because we go through the gateway any request to `/` is picked by the gateway as a request to it. This is why the root request test fails.

It's not really adding any value and was added more for demonstration purposes. So the simplest solution is just to remove it.