Remove newman html reporter
Having added it we got 2 vulnerability reports. A check of the project shows it has not been updated in more than a year. When we ran it we confirmed that the output was not as expected. A quick check found [newman-reporter-htmlextra](https://github.com/DannyDainton/newman-reporter-htmlextra). This is much more actively maintained and matches what we have previously seen when viewing a HTML version of the newman output.