Implement htmlextra helper
Key driver for these changes was still allowing users of the project to request the `html` reporter but us to actually use the `htmlextra` reporter.

As we needed to code a method to handle this we also added some logic to deal with the case of the args and remove any duplicates. Doing this both helps us replacing `html` with `htmlextra` whilst also cleaning things up before passing them to newman.