Add new html reporter (#5)
The HTML reporter whilst maintained by Postman is not provided as a built-in reporter to Newman. You have to add it as an additional package.

However, when we reviewed it we found

- it had not been updated for over a year
- contained dependencies with vulnerabilities
- offered far less functionality than community made alternatives

So we found [newman-reporter-htmlextra](https://github.com/DannyDainton/newman-reporter-htmlextra). This is much more actively maintained and matches what we have previously seen when viewing a HTML version of the Newman output.

This change covers everything required to implement it as part of the project.