Add support for specifying folders on CLI
You have to specify a collection when using Postman and Newman. However, if your collection contains a folder of requests, you can tell Newman to just one that one folder. You can also pass it an array of folders.

Using folders is really the only means we have to group requests into a 'test'. So by adding support to specify folders in our CLI options, we are effectively adding support for declaring which test(s) you wish to run.