Add transaction performance test
https://trello.com/c/gTcQmR8M

This adds a test (folder) to the Postman collection which creates a new bill run and then adds 100 transactions to it. The test includes logic that ensures we have a mix of customers, financial years and licence numbers. It has been created to give an early indicator for how changes to how a transaction is created impacts average response times.