Remove transaction performance test
The `transaction-performance-test` was great whilst we were re-developing creating a transaction in V2 of the API. We could see at each stage what impact our changes were having on performance.

Other than fixes, our work is now done. Plus thanks to @anwarisse1 efforts we can do something similar in our JMeter tests. There is also issues with keeping the test in this project

- you can't just 'run the collection' as the performance test will run as well
- the need to set next request to `null` in the performance test kills any tests that follows it

We were initially going to look at adding some new config to automatically exclude the performance test. But upon review, there is no real need to keep it so it'd be cleaner and easier to just remove it.