Update duplicate ClientId create transaction test
Currently request `4. 409 Conflict Error is returned when creating a transaction with Client ID thats already in use` in `create-pre-sroc-transaction` is checking that the status code is 409 when you attempt to create a transaction with a duplicate `clientId`.

It also has 2 additional checks

- **New Transaction ID has not been generated**
- **New Client ID has not been generated**

The names of the tests read correct but the implementation is checking for the opposite, for example, `pm.expect(response.id).to.eql(environment.transaction_id);`. That's checking that the response _does include_ an ID.

This change updates the implementation of the checks to match the test descriptions. It also adds a new check that the error message is as expected.