Update region mandatory error message check
The `25. "region" is mandatory and Transaction is not created` in `create-pre-sroc-transaction` helped us spot that we were not validating the region correctly.

See https://github.com/DEFRA/sroc-charging-module-api/pull/185

But when we came to retest we found the region check was expecting a different error message to what we actually return. This change updates the check to match what the API will actually return.