Update newLicence to be subjectToMinimumCharge
We know that the definition of what invoices should be considered for minimum charge is changing to not just being those for 'new licences'. So the development team have got ahead of the change and updated the property in the request to use a more generic name in Version 2 of the API.

So, it will now expect the property `subjectToMinimumCharge` to be a boolean set to `true` or `false`.

Obviously, this means means current tests that rely on `newLicence` being set need to be updated. This change does that.

> Note: `subjectToMinimumCharge` is the development team's current pitch for what it should be. And if agreed, then job done! If however a different term is preferred it will be an extremely small change to update the API. This is part of why we have gone ahead with it now.