Fix admin regime request tests
If you try and run the requests in the admin folder you get an error. This is because the `Show regimes` request still assumes we are using the v1 format of regime 'slug'.

This fixes the requests so we can get closer to a point where you can run everything in the collection and not have to stipulate specific folders.