Update validation msg asserts / housekeeping! (#14)
* Update validation msg asserts / housekeeping!

With access to the automated tests the development team have now been able to run them and start getting feedback for themselves. They have already helped to expose 2 other issues that need fixing!

At the same time, they did spot some other things that this change aims to rectify

- period start/end date validation messages have been updated to match what will actually be sent back
- fixed typo in some test code
- removed a duplicated standalone `calculate-standalone-pre-sroc-charge` request
- fixed the numbering in `calculate-standalone-pre-sroc-charge`

* Updated postman collection with changes

First imported the collection, made the changes, then exported it back to the project. Some changes will just be as a result of this process.