Update newLicence to be subjectToMinimumCharge (#18)
* Update validation msg asserts / housekeeping!

With access to the automated tests the development team have now been able to run them and start getting feedback for themselves. They have already helped to expose 2 other issues that need fixing!

At the same time, they did spot some other things that this change aims to rectify

- period start/end date validation messages have been updated to match what will actually be sent back
- fixed typo in some test code
- removed a duplicated standalone `calculate-standalone-pre-sroc-charge` request
- fixed the numbering in `calculate-standalone-pre-sroc-charge`

* Updated postman collection with changes

First imported the collection, made the changes, then exported it back to the project. Some changes will just be as a result of this process.

* Remove transaction performance test

The `transaction-performance-test` was great whilst we were re-developing creating a transaction in V2 of the API. We could see at each stage what impact our changes were having on performance.

Other than fixes, our work is now done. Plus thanks to @anwarisse1 efforts we can do something similar in our JMeter tests. There is also issues with keeping the test in this project

- you can't just 'run the collection' as the performance test will run as well
- the need to set next request to `null` in the performance test kills any tests that follows it

We were initially going to look at adding some new config to automatically exclude the performance test. But upon review, there is no real need to keep it so it'd be cleaner and easier to just remove it.

* Remove transaction-performance folder

* Update duplicate ClientId create transaction test

Currently request `4. 409 Conflict Error is returned when creating a transaction with Client ID thats already in use` in `create-pre-sroc-transaction` is checking that the status code is 409 when you attempt to create a transaction with a duplicate `clientId`.

It also has 2 additional checks

- **New Transaction ID has not been generated**
- **New Client ID has not been generated**

The names of the tests read correct but the implementation is checking for the opposite, for example, `pm.expect(response.id).to.eql(environment.transaction_id);`. That's checking that the response _does include_ an ID.

This change updates the implementation of the checks to match the test descriptions. It also adds a new check that the error message is as expected.

* Update checks and add new one

* Update region mandatory error message check

The `25. "region" is mandatory and Transaction is not created` in `create-pre-sroc-transaction` helped us spot that we were not validating the region correctly.

See https://github.com/DEFRA/sroc-charging-module-api/pull/185

But when we came to retest we found the region check was expecting a different error message to what we actually return. This change updates the check to match what the API will actually return.

* Update message to be checked for

* Update newLicence to be subjectToMinimumCharge

We know that the definition of what invoices should be considered for minimum charge is changing to not just being those for 'new licences'. So the development team have got ahead of the change and updated the property in the request to use a more generic name in Version 2 of the API.

So, it will now expect the property `subjectToMinimumCharge` to be a boolean set to `true` or `false`.

Obviously, this means means current tests that rely on `newLicence` being set need to be updated. This change does that.

> Note: `subjectToMinimumCharge` is the development team's current pitch for what it should be. And if agreed, then job done! If however a different term is preferred it will be an extremely small change to update the API. This is part of why we have gone ahead with it now.

* Apply change to request and test