Update section126Factor decimal places error check (#20)
* Update section126Factor decimal places error check

Both `calculate-pre-sroc-charge-in-transaction` and `calculate-standalone-pre-sroc-charge` have checks that ensure the `section126Factor` of a charge request does not allow more than 3 decimal places. The transaction file the value gets exported to has this limit and we do not want to take responsibility for rounding values sent by client systems.

So, we must reject `section126Factor` values which have more than 3 decimal places. We have just [fixed the API](https://github.com/DEFRA/sroc-charging-module-api/pull/191) to ensure this is the case. The only thing we need to change in the tests is the check for the error message (it's not quite the same).

* Update error message checks