Add tests for string case matching in calc. charge
Whilst testing the `/generate` bill run endpoint we came across a situation where the `source`, `season`, and `loss` values were found to be invalid. Yet the rules service still returned a 'valid' result. Hidden away in the response was a message telling us that the values we had sent through did not match those in the rules service.

After some investigation we found it was purely the 'case' of the strings that were used causing the problem. The values were fine.

So, we are doing 2 things

- we are updating the [SROC Charging Module API](https://github.com/defra/sroc-charging-module-api) to automatically 'title case' any values sent through
- we are also updating it to check for messages from the rules service and respond appropriately

To support these changes this adds some tests to check

- the automatic title casing of strings
- that we get an error response from the CMA if the rules service reports issues with the values